requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
behave==1.2.6
certifi==2023.5.7
cffi==1.15.1
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
parse==1.19.1
parse-type==0.6.2
pi==0.1.2
pluggy==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0